Increased Font Size + more
I increased the font size of the timer and deleted unnecessary code that I commented out. I changed how the stopwatch works. Previously, the stopwatch would just decrement the seconds variable, which would be declared outside of the function. This would allow for errors that would be a pain to fix. This is why I got the idea to define the dates every second, that way it changes every time the function is run. TLDR: Made code easier to work with.